devcor3/m2/requirements.txt
requests
pyyaml
pyang
